Implements electricity grid system
Introduces a new electricity grid system based on Circuits.

This system allows for the connection of different electrical
components such as coils, generators, and engines.

The system also supports adding and removing electrical structure
and merging circuits.

This commit also updates the engine and generator classes to
include position data.