Adds in-game world saving and loading
Implements world saving and loading functionality using JSON serialization.

This allows the game to persist world state, including tilemap and structure data, to the persistent data path, and to load it back.
Introduces a new `WorldLoaderUI` for listing save files and loading them.
Also introduces a `PlayerController` which allow to select a structure and place it on the map